+++ .specify/plugins/tdk-utils/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this plugin will be documented in this file.
 
 Format: [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), Semver.
 
+## [3.0.3] - 2026-07-29
+
+### Changed
+- tdk-scout documents the file ceiling, the oversize failure mode, and the --include/--ignore narrowing flags
+
 ## [3.0.2] - 2026-07-28
 
 ### Changed